Updating the flags of Switzerland and Vatican City
Reference for this suggestion:
https://forums.online-go.com/t/square-flags-for-vatican-city-and-switzerland/38957

The Swiss flag and the Vatican City flag are square sovereign-state
flags.
Representing them with a rectangular flag is incorrect.
This commit adds these two flags with a square shape.
The css and scss files are updated to match the correct new position.

The old flags were not removed from the png files as no tool is
available to do it and I wanted to avoid doing it by hand with an
external tool and potentially mess up the flags16.png and flags32.png
files.

A mention regarding the lack of gloss on these flags:
I couldn't generate correctly the gloss with the add_gloss.py script.
The script executes fine with no error but the output is an image of
the same size as the one in input and the render is a fading white
gradient from the top to the bottom over a black background.
It seems the script fails to use the input image as the background.